* Source/NSBitmapImageRep.m: * Source/NSColor.m: * Source/GSThemeTools.m: * Source/NSGraphicsContext.m: * Headers/AppKit/NSColor.h: * Headers/AppKit/NSGraphicsContext.h: * ColorPickers/GSGrayColorPicker.m: * ColorPickers/GSCMYKColorPicker.m: * ColorPickers/GSWheelColorPicker.m: * ColorPickers/GSRGBColorPicker.m: * ColorPickers/GSHSBColorPicker.m: Migrate float to CGFloat in NSColor
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@34070 72102866-910b-0410-8b05-ffd578937521

+++ ColorPickers/GSCMYKColorPicker.m
@@ -64,7 +64,7 @@
 
 - (void)setColor:(NSColor *)color
 {
-  float cyan, magenta, yellow, black, alpha;
+  CGFloat cyan, magenta, yellow, black, alpha;
   NSColor *c;
 
   if (updating)
