+++ ColorPickers/GSHSBColorPicker.m
@@ -62,7 +62,7 @@
 
 - (void)setColor:(NSColor *)color
 {
-  float hue, saturation, brightness, alpha;
+  CGFloat hue, saturation, brightness, alpha;
   NSColor *c;
 
   if (updating)
